+++ src/errors.js
@@ -2,6 +2,7 @@ const { SCRAPE_TYPES, SEARCH_TYPES } = require('./consts');
 
 module.exports = {
     redirectedToLogin: () => new Error('Page got redirected into login page. Instagram is blocking access. Retrying with different IP and browser. Make sure you are accessing public profile or being logged in'),
+    challengePage: () => new Error('Challenge page appeared, will be retried but most likely won\'t work'),
     unsupportedPage: () => new Error('This instagram page is not supported.'),
     proxyIsRequired: () => new Error('Proxy is required to run this actor'),
     urlsAreRequired: () => new Error('Please provide urls configuration'),